requirements.txt
altgraph==0.17.4
attrs==24.2.0
cattrs==23.2.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coremltools==7.2
et-xmlfile==1.1.0
filelock==3.15.4
fsspec==2024.6.1
future==1.0.0
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kraken==2.0.8
lxml==5.3.0
macholib==1.16.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.0
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pillow==10.4.0
protobuf==3.20.3
pyaml==24.7.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyPDF2==3.0.1
pypdfium2==4.30.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzbar==0.1.9
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rpds-py==0.20.0
scipy==1.14.1
setuptools==73.0.1
six==1.16.0
sympy==1.13.2
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
